added costs to the captured data. refactored consumption to double

+++ src/main/kotlin/com/bullit/energysimulator/Domain.kt
@@ -4,18 +4,18 @@ import java.time.LocalDateTime
 
 sealed interface Consumption {
     val dateTime: LocalDateTime
-    val amountConsumed: Long
+    val amountConsumed: Double
 }
 
 data class PowerConsumption(
     override val dateTime: LocalDateTime,
-    override val amountConsumed: Long,
+    override val amountConsumed: Double,
     val rate: Rate
 ): Consumption
 
 data class GasConsumption(
     override val dateTime: LocalDateTime,
-    override val amountConsumed: Long
+    override val amountConsumed: Double
 ): Consumption
 
 interface RawCSVData {
@@ -24,13 +24,13 @@ interface RawCSVData {
 
 data class RawCSVDataPower(
     override val dateTime: LocalDateTime,
-    val t1: Long,
-    val t2: Long
+    val t1: Double,
+    val t2: Double
 ): RawCSVData
 
 data class RawCSVDataGas(
     override val dateTime: LocalDateTime,
-    val meterReading: Long
+    val meterReading: Double
 ): RawCSVData
 
 enum class Rate {
